requirements.txt
fastapi==0.87.0
uvicorn[standard]==0.19.0
sqlalchemy==1.4.44
alembic==1.8.1
psycopg2-binary==2.9.5
flake8
coverage
python-dotenv
